Renamed module from 'client' to 'web'
