+++ graphql_client_web/src/lib.rs
@@ -2,5 +2,5 @@
     note = "graphql_client_web is deprecated. The web client is now part of the graphql_client crate, with the default \"client\" feature."
 )]
 
-pub use graphql_client::client::*;
+pub use graphql_client::http::*;
 pub use graphql_client::{self, *};
